Gracefully handle relay host timeout, fix typos, and move SMBHashCapture location
