+++ lib/msf/core/exploit/remote/smb/relay/ntlm/server_client.rb
@@ -36,10 +36,12 @@ module Msf::Exploit::Remote::SMB::Relay::NTLM
       )
 
       if relayed_connection.nil?
-        @relay_targets.on_relay_end(target, identity: session.metadata[:identity], is_success: false)
+        @relay_targets.on_relay_end(session.metadata[:relay_target], identity: session.metadata[:identity], is_success: false)
+        session.metadata[:relay_mode] = false
+      else
+        session.metadata[:relay_mode] = true
       end
 
-      session.metadata[:relay_mode] = true
       session.metadata[:relayed_connection] = relayed_connection
       session.state = :in_progress
