+++ modules/exploits/windows/smb/smb_relay.rb
@@ -19,7 +19,7 @@ class MetasploitModule < Msf::Exploit::Remote
   Rank = ExcellentRanking
   include ::Msf::Exploit::Remote::SocketServer
 
-  include ::Msf::Exploit::SMBHashCapture
+  include ::Msf::Exploit::Remote::SMB::Server::HashCapture
   include ::Msf::Exploit::Remote::SMB::Client::Psexec
   include ::Msf::Exploit::Powershell
   include Msf::Exploit::EXE
@@ -118,7 +118,7 @@ class MetasploitModule < Msf::Exploit::Remote
       [
         OptString.new('SMBSHARE', [false, 'The share to connect to, can be an admin share (ADMIN$,C$,...) or a normal read/write folder share', ''], aliases: ['SHARE']),
         OptAddressRange.new('RELAY_TARGETS', [true, 'Target address range or CIDR identifier to relay to'], aliases: ['SMBHOST']),
-        OptAddressLocal.new('SRVHOST', [true, 'The local host or network interface to listen on. This must be an address on the local machine or 0.0.0.0 to listen on all addresses.']),
+        OptAddress.new('SRVHOST', [ true, 'The local host to listen on.', '0.0.0.0' ]),
         OptPort.new('SRVPORT', [true, 'The local port to listen on.', 445]),
         OptString.new('CAINPWFILE', [false, 'Name of file to store Cain&Abel hashes in. Only supports NTLMv1 hashes. Can be a path.', nil]),
         OptString.new('JOHNPWFILE', [false, 'Name of file to store JohnTheRipper hashes in. Supports NTLMv1 and NTLMv2 hashes, each of which is stored in separate files. Can also be a path.', nil]),
@@ -258,7 +258,7 @@ class MetasploitModule < Msf::Exploit::Remote
     def start
       @listener_sock = Rex::Socket::TcpServer.create(sock_options)
       @listener_server = Msf::Exploit::Remote::SMB::Relay::NTLM::Server.new(**smb_server_options(@listener_sock))
-      @listener_thread = Rex::ThreadFactory.spawn('UDPLDAPServerListener', false) do
+      @listener_thread = Rex::ThreadFactory.spawn('SMBRelayServerListener', false) do
         @listener_server.run
       rescue StandardError => e
         elog(e)
@@ -270,7 +270,7 @@ class MetasploitModule < Msf::Exploit::Remote
         @listener_server.close if @server && !@server.closed?
         @listener_thread.kill if @listener_thread
       rescue StandardError => e
-        print_error('Failed closing SMV server')
+        print_error('Failed closing SMB server')
         elog('Failed closing SMB server', error: e)
       end
